requirements.txt
isodate
requests
mysql-connector-python
